+++ docs/release-notes.md
@@ -9,7 +9,7 @@ title: jqwik Release Notes
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 ### Table of Contents  
 
-- [1.5.6-SNAPSHOT](#156-snapshot)
+- [1.5.6](#156)
 - [1.5.5](#155)
 - [1.5.4](#154)
 - [1.5.3](#153)
@@ -27,7 +27,13 @@ title: jqwik Release Notes
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
 
-## 1.5.6-SNAPSHOT
+## 1.5.6
+
+<p style="padding-left:1em;font-size:larger">
+<a href="/docs/1.5.6/user-guide.html">User guide</a>
+and
+<a href="/docs/1.5.6/javadoc/index.html">javadoc</a>
+</p>
 
 #### New and Enhanced Features